app/requirements.txt
Flask==3.1.0
boto3==1.36.13
python-dotenv==1.0.1
